move ConfusionMatrix into its own file
